Implement storage migration for operator content (#556)
